chore: move misc methods to miscFunctions class
