+++ src/main/java/com/limechain/runtime/hostapi/functions/MiscellaneousHostFunctions.java
@@ -2,14 +2,18 @@ package com.limechain.runtime.hostapi.functions;
 
 import com.limechain.runtime.hostapi.HostApi;
 import lombok.experimental.UtilityClass;
+import lombok.extern.java.Log;
+import org.apache.tomcat.util.buf.HexUtils;
 import org.wasmer.ImportObject;
 import org.wasmer.Type;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.logging.Level;
 
 @UtilityClass
-public class MiscellaneousHostFunctions {
+@Log
+public class MiscellaneousHostFunctions extends HostApi {
 
     public static List<ImportObject> getFunctions() {
         return Arrays.asList(
@@ -20,21 +24,47 @@ public class MiscellaneousHostFunctions {
 
                 }, List.of(Type.I64)),
                 HostFunctions.getImportObject("ext_misc_print_hex_version_1", argv -> {
-                    HostApi.extMiscPrintHex((long) argv.get(0));
+                    extMiscPrintHex((long) argv.get(0));
                     }, List.of(Type.I64)),
                 HostFunctions.getImportObject("ext_misc_runtime_version_version_1", argv -> {
                     return argv;
                 }, List.of(Type.I64), Type.I64),
                 HostFunctions.getImportObject("ext_logging_log_version_1", argv -> {
-                    HostApi.extLoggingLog((Integer) argv.get(0), (Long) argv.get(1), (Long) argv.get(2));
+                    extLoggingLog((Integer) argv.get(0), (Long) argv.get(1), (Long) argv.get(2));
                 }, Arrays.asList(Type.I32, Type.I64, Type.I64)),
                 HostFunctions.getImportObject("ext_logging_max_level", argv -> {
-                    //TODO: Update
                     return argv;
                 }, List.of(), Type.I64),
                 HostFunctions.getImportObject("ext_panic_handler_abort_on_panic_version_1", argv -> {
-                    HostApi.extPanicHandlerAbortOnPanicVersion1((Long) argv.get(0));
+                    extPanicHandlerAbortOnPanicVersion1((Long) argv.get(0));
                 }, List.of(Type.I64)));
     }
 
+    private static void extMiscPrintHex(long pointer) {
+        byte[] data = getDataFromMemory(pointer);
+        log.info(HexUtils.toHexString(data));
+    }
+
+    private static void extLoggingLog(int level, long targetPtr, long messagePtr) {
+        byte[] target = getDataFromMemory(targetPtr);
+        byte[] message = getDataFromMemory(messagePtr);
+
+        log.log(getLogLevel(level), new String(target) + ": " + new String(message));
+    }
+
+
+    private static void extPanicHandlerAbortOnPanicVersion1(long messagePtr) {
+        byte[] data = getDataFromMemory(messagePtr);
+        log.severe(String.valueOf(data));
+    }
+
+    private static Level getLogLevel(int i) {
+        return switch (i) {
+            case 0 -> Level.SEVERE;
+            case 1 -> Level.WARNING;
+            case 2 -> Level.INFO;
+            case 3 -> Level.FINE;
+            default -> Level.FINEST;
+        };
+    }
 }
